+++ src/Observer/CustomerAddressSaveObserver.php
@@ -34,6 +34,7 @@ class CustomerAddressSaveObserver implements ObserverInterface
     {
         /** @var \Magento\Customer\Model\Address $address */
         $address = $observer->getEvent()->getCustomerAddress();
+        $houseNumberFull = '';
 
         // convert old address to new format
         $streetParts = $this->addressHelper->getMultiLineStreetParts($address->getStreet());
@@ -41,10 +42,25 @@ class CustomerAddressSaveObserver implements ObserverInterface
             // Get street, house number, etc from line 1
             $streetParts = $this->addressHelper->getStreetParts($address->getStreet());
         }
+        if ($address->getStreetName() != '') {
+            $streetParts['street'] = $address->getStreetName();
+        }
+        if ($address->getHouseNumber() != '') {
+            $streetParts['house_number'] = $address->getHouseNumber();
+            $houseNumberFull = $streetParts['house_number'];
+        }
+        if ($address->getHouseNumberAddition() != '') {
+            $streetParts['addition'] = $address->getHouseNumberAddition();
+            $houseNumberFull .= ' ' . $streetParts['addition'];
+        }
 
         // @todo: check if already has values for house_number, etc
         $address->setStreetName($streetParts['street']);
         $address->setHouseNumber($streetParts['house_number']);
         $address->setHouseNumberAddition($streetParts['addition']);
+
+
+
+        $address->setStreet($streetParts['street'] . " " . $houseNumberFull);
     }
 }