+++ src/Setup/InstallData.php
@@ -145,13 +145,12 @@ class InstallData implements InstallDataInterface
                     'type'             => 'varchar',
                     'label'            => 'Street Name',
                     'input'            => 'text',
-                    'required'         => false,
+                    'required'         => true,
                     'visible'          => true,
                     'visible_on_front' => true,
                     'user_defined'     => true,
-                    'position'       => 74,
+                    'position'       => 76,
                     'system'           => 0,
-                    'input_filter'     => 'stripTags'
                 ]
             );
         }
@@ -186,9 +185,8 @@ class InstallData implements InstallDataInterface
                     'visible'          => true,
                     'visible_on_front' => true,
                     'user_defined'     => true,
-                    'position'       => 75,
+                    'position'       => 74,
                     'system'           => 0,
-                    'input_filter'     => 'stripTags'
                 ]
             );
         }
@@ -222,9 +220,8 @@ class InstallData implements InstallDataInterface
                     'visible'          => true,
                     'visible_on_front' => true,
                     'user_defined'     => true,
-                    'position'       => 76,
+                    'position'       => 75,
                     'system'           => 0,
-                    'input_filter'     => 'stripTags'
                 ]
             );
         }